requirements.txt
Django>=5.2,<7.0
djangorestframework>=3.15
django-filter>=24.0
drf-spectacular>=0.27
python-dateutil>=2.9
django-htmx>=1.17
python-dotenv>=1.0
mcp>=1.0
